+++ views/admin/settings.php
@@ -145,6 +145,15 @@
                             </p>
                         </div>
                     </div>
+
+                    <div class="uk-form-row">
+                        <label class="uk-form-label">{{ 'Image quality' | trans }}</label>
+                        <div class="uk-form-controls uk-form-controls-text">
+                            <p class="uk-form-controls-condensed">
+                                <input type="number" v-model="config.images.image_quality" class="uk-form-width-small">
+                            </p>
+                        </div>
+                    </div>
                 </li>
             </ul>